+++ dubbo-spring-boot-actuator/src/main/resources/META-INF/dubbo-web-endpoins-default.properties
@@ -1,0 +1,22 @@
+# Dubbo Web Endpoints Default Properties is loaded by @PropertySource with low order,
+# those values of properties can be override by higher PropertySource
+# @see DubboEndpointsAutoConfiguration
+
+# "management.endpoints.web.base-path" should not be configured in this file
+
+# Set enabled for Dubbo Web Endpoints
+management.endpoint.dubbo.enabled = true
+management.endpoint.dubbo-shutdown.enabled = false
+management.endpoint.dubbo-configs.enabled = true
+management.endpoint.dubbo-services.enabled = false
+management.endpoint.dubbo-references.enabled = false
+management.endpoint.dubbo-properties.enabled = true
+
+# Re-defines path-mapping of Dubbo Web Endpoints
+management.endpoints.web.path-mapping.dubbo-shutdown = dubbo/shutdown
+management.endpoints.web.path-mapping.dubbo-configs = dubbo/configs
+management.endpoints.web.path-mapping.dubbo-services = dubbo/services
+management.endpoints.web.path-mapping.dubbo-references = dubbo/references
+management.endpoints.web.path-mapping.dubbo-properties = dubbo/properties
+
+
