+++ dubbo-spring-boot-actuator/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.alibaba.boot.dubbo.actuate.autoconfigure.DubboEndpointAutoConfiguration,\
+com.alibaba.boot.dubbo.actuate.autoconfigure.DubboEndpointsAutoConfiguration,\
 com.alibaba.boot.dubbo.actuate.autoconfigure.DubboHealthIndicatorAutoConfiguration
-
-org.springframework.boot.actuate.autoconfigure.ManagementContextConfiguration=\
-com.alibaba.boot.dubbo.actuate.autoconfigure.DubboMvcEndpointManagementContextConfiguration
