+++ dubbo-spring-boot-samples/dubbo-spring-boot-sample-consumer/src/main/resources/application.properties
@@ -1,8 +1,7 @@
 # Spring boot application
 spring.application.name = dubbo-consumer-demo
 server.port = 8080
-management.port = 8081
-management.security.enabled = false
+management.server.port = 8081
 
 # Dubbo Config properties
 ## ApplicationConfig Bean
@@ -23,3 +22,14 @@ endpoints.dubbo.enabled = true
 # Dubbo Health
 ## StatusChecker Name defaults (default : "memory", "load" )
 management.health.dubbo.status.defaults = memory
+
+# Enables Dubbo All Endpoints
+management.endpoint.dubbo.enabled = true
+management.endpoint.dubbo-shutdown.enabled = true
+management.endpoint.dubbo-configs.enabled = true
+management.endpoint.dubbo-services.enabled = true
+management.endpoint.dubbo-references.enabled = true
+management.endpoint.dubbo-properties.enabled = true
+
+# Exposes all web endpoints
+management.endpoints.web.exposure.include = *
