+++ dubbo-spring-boot-samples/dubbo-spring-boot-sample-provider/src/main/resources/application.properties
@@ -1,8 +1,6 @@
 # Spring boot application
 spring.application.name = dubbo-provider-demo
 server.port = 9090
-management.port = 9091
-management.security.enabled = false
 
 # Base packages to scan Dubbo Components (e.g @Service , @Reference)
 dubbo.scan.basePackages  = com.alibaba.boot.dubbo.demo.provider.service
